refact: card interfaces
