+++ src/types/index.ts
@@ -1,4 +1,5 @@
 export * from './user'
 export * from './todo'
+export * from './card'
 export * from './product'
 export * from './wheater-forecast'
